update and destroy user

+++ app/controllers/users_controller.rb
@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
   before_filter :authenticate_user!
-
+  after_action :verify_authorized
 
   def index
     @users = User.all
@@ -14,8 +14,26 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
+    authorize user
     user.destroy
     redirect_to users_path, :notice => "user deleted"
   end
+
+  def update
+    @user = User.find(params[:id])
+    authorize @user
+
+    if @user.update_attributes(secure_params)
+      redirect_to users_path, :success => "User updated"
+    else
+      redirect_to user_path, :alert => "Unable to update"
+    end
+  end
+
+  private
+
+    def secure_params
+      params.require(:user).permit(:role)
+    end
 
 end
